backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
openai>=1.30.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.5.0
sqlalchemy>=2.0.30
